pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = ""
authors = [
    {name = "koshishshresthaa",email = "koshishshrestha74@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "cloudscraper (>=1.2.71,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "crontab (>=1.0.1,<2.0.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
